Added functionality for searching all columns with 1 method in CharacterRequests
