+++ .idea/dataSources/7e582771-a0e3-4802-8a45-0a2cc3149532/storage_v2/_src_/schema/information_schema.FNRwLQ.meta
@@ -1,0 +1,2 @@
+#n:information_schema
+!<md> [null, 0, null, null, -2147483648, -2147483648]
